fix(public): add move apis methods
