+++ clients/packages/webpage-client/src/pages/api/data/targets.ts
@@ -1,6 +1,6 @@
 import gql from 'graphql-tag';
 
-import { client } from '../../../graphql-app';
+import { client } from '../../../apis/graphql';
 
 const query = gql`
   query Targets($widgetId: Int!) {
