+++ clients/packages/webpage-client/src/pages/index.tsx
@@ -3,9 +3,9 @@ import getConfig from 'next/config';
 import i18n from "i18next";
 import { I18nextProvider, initReactI18next } from "react-i18next";
 import Error404 from './404';
-import asyncFilterBlocksGraphql from '../graphql-app/filterBlocks';
-import asyncFilterMobilizationsGraphql from '../graphql-app/filterMobilizations';
-import asyncFilterWidgetsGraphql from '../graphql-app/filterWidgets';
+import asyncFilterBlocksGraphql from '../apis/graphql/filterBlocks';
+import asyncFilterMobilizationsGraphql from '../apis/graphql/filterMobilizations';
+import asyncFilterWidgetsGraphql from '../apis/graphql/filterWidgets';
 import MeuRioStyles from '../components/MeuRioStyles';
 import Styles from '../bonde-webpage/Styles';
 import MobilizationConnected from '../components/MobilizationConnected';
